// src/types/user.ts
interface UserdataModel {
  id: string | null
  nama: string | null
  username: string | null
  no_hp: string | null
  level_id: string | null
  active: string | null
  dapil_id: string | null
  kecamatan_id: string | null
  kelurahan_id: string | null
  tps_id: string | null
  created_at: string | null
  creator_id: string | null
  update_at: string | null
  updated_id: string | null
  dapil: string | null
  kecamatan: string | null
  kelurahan: string | null
  tps: string | null
  kabupaten: string | null
  upload_photo: boolean| null
}

export type { UserdataModel }